web/requirements.txt
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
cycler==0.10.0
filelock==3.0.12
Flask==1.1.2
Flask-Admin==1.5.7
Flask-Script==2.0.6
gdown==3.12.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.3
numpy==1.19.4
pandas==1.1.5
Pillow==8.0.1
pycairo @ git+https://github.com/pygobject/pycairo.git@dad2b743b957c9579542e5923a8e5aa5ca790b88
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.4
requests==2.25.0
six==1.15.0
torch==1.7.1
torchaudio==0.7.2
torchvision==0.8.2
tqdm==4.54.1
typing-extensions==3.7.4.3
urllib3==1.26.2
Werkzeug==1.0.1
WTForms==2.3.3
